Move action handlers to deployment package
