+++ internal/deployment/action_finished_adapter.go
@@ -1,4 +1,4 @@
-package problem
+package deployment
 
 import (
 	cloudevents "github.com/cloudevents/sdk-go/v2"
