+++ internal/deployment/action_triggered_event_handler.go
@@ -1,4 +1,4 @@
-package problem
+package deployment
 
 import (
 	"context"
